apimachinery: decompose meta/v1/unstructured in order to cut import cycle
