+++ staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured/BUILD
@@ -17,11 +17,16 @@ go_test(
 
 go_library(
     name = "go_default_library",
-    srcs = ["unstructured.go"],
+    srcs = [
+        "doc.go",
+        "unstructured.go",
+        "zz_generated.deepcopy.go",
+    ],
     tags = ["automanaged"],
     deps = [
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
